Smartemailorganizer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
scikit-learn==1.3.2
pyjwt==2.8.0
bcrypt==4.1.1
email-validator==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
